Avoid blocked event bus (#7)
The old synchronization of asynchronous non-concurrent event handlers could lead to a blocked event bus with a single long-running event handler. For this to happen, the following situation had to occur:

A event had to be thrown and a (asynchronous non-concurrent) handler has to take some time to process this event
During this processing more events have to be thrown that have to be handled by this blocked event handler.
In this case, the handling of the following events are blocked by the synchronization in the SynchronizedEventHandler class until the first event has been processed completely. Because the synchronization happens in the handleEvent method inside the handler, this blocks a thread of the executor per event, until all threads are blocked an therefore no more asynchronous events whatsoever can be processed any more.

To avoid such situations, the synchronization is now done outside of the executor and its threads. To do so, the new class ExecutorSerializer is introduced. This class can be seen as a steward, that blocks events for event handlers, whenever a event handler of the same type is already queued in the executor to process another event. The waiting event is put in a queue outside the executor an will be taken into account again, whenever the first process has finished.